+++ test/losses/test_losses.py
@@ -45,6 +45,25 @@ class TestBinaryFocalLossWithLogits:
             == ()
         )
 
+    @pytest.mark.parametrize("pos_weight", [None, 1, 5])
+    def test_smoke_pos_weight(self, device, dtype, pos_weight):
+        logits = torch.rand(2, 3, 2, dtype=dtype, device=device)
+        labels = torch.rand(2, 3, 2, dtype=dtype, device=device)
+
+        assert (
+            kornia.losses.binary_focal_loss_with_logits(
+                logits,
+                labels,
+                alpha=0.5,
+                gamma=2.0,
+                reduction="mean",
+                pos_weight=None
+                if pos_weight is None
+                else torch.full([logits.size(-1)], pos_weight, dtype=dtype, device=device),
+            ).shape
+            == ()
+        )
+
     def test_jit(self, device, dtype):
         logits = torch.rand(2, 3, 2, dtype=dtype, device=device)
         labels = torch.rand(2, 3, 2, dtype=dtype, device=device)